Use the biClrUsed field for the palette size in SDL_LoadBMP
While it sounds like some .bmp files have an invalid or corrupted value
here, there also exists some code (such as SDL 1.2's testpalette) which
rely on this being propagated to the SDL_Palette's ncolors field.
Otherwise, we end up with too big a palette (potentially with garbage
entries), which -- in the case of testpalette -- causes memory
corruption due to overflowing a buffer.

The original code here seems to have come from SDL_Image, so there's
always the possibility for people to use that if they need to load
.bmp files with invalid biClrUsed fields, I guess...