wayland: Support building with static wayland/libdecor (fix #4543)
When wayland is not dynamically linked (--enable-wayland-shared=no),
libdecor is always also statically linked (it's treated as part of
Wayland in SDL_waylanddyn.{c,h} and SDL_waylandsym.h). However, it
doesn't include libdecor.h unless SDL_VIDEO_DRIVER_WAYLAND_DYNAMIC is
set, so it fails to build at all unless Wayland is dynamically loaded.

We can't simply move the libdecor.h include above the #ifdef
SDL_VIDEO_DRIVER_WAYLAND_DYNAMIC block, as libdecor.h itself #includes
wayland headers we need to replace with #defines. Instead, we duplicate
the #include for the statically linked case.

Note that this doesn't fix any of the underlying issues of libdecor
being treated as part of wayland, it just fixes the build. A better
solution would probably be to decouple the wayland dynamic loading from
the libdecor dynamic loading completely, though that is a lot more work…