wayland: Wayland cursors should use premultiplied alpha
It turns out that Wayland's WL_SHM_FORMAT_ARGB8888 format (and, indeed,
all wayland RGBA formats) should be treated as premultiplied. SDL
surfaces tend not to be premultiplied, and this is assumed by other
backends when dealing with cursors.

This change premultiplies the cursor surface in Wayland_CreateCursor().
A helper function -- premultiply_surface_alpha_ARGB8888 -- is created to
do the conversion, based somewhat on the e similar
legacy_alpha_premultiply_ARGB8888 function in KMSDRM, which has the same
requirement to have premultiplied alpha for the cursor.

This should fix #4856